improve test reliability
